requirements.txt
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Flask==3.0.3
gcloud==0.18.3
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jws==0.1.3
MarkupSafe==3.0.2
msgpack==1.1.0
oauth2client==4.1.3
proto-plus==1.24.0
protobuf==5.29.0rc1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.9.0
pyparsing==3.2.0
Pyrebase4==4.8.0
python-dotenv==1.0.1
python-jwt==2.0.1
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
setuptools==75.2.0
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.20
Werkzeug==3.0.4
